requirements.txt
pretty_midi
keras
gym
tensorflow-gpu
joblib
tqdm
python-midi
